# Increase performance of YCbCr and RGBA encoding