@Kagami 2022-07-04
1st and 3rd cases seem to be the same?
Also, I don't particularly like that the whole loop is duplicated...

For best performance both decoding and colorformat conversion should be done with SIMD and passed as raw bytes to the AV1 encoding library. Though AVIF encoding is so slow and it doesn't matter that much.